Rename file to singular naming as kubeconform does not find the plural one
We added a CustomResourceDefinition to our repository which broke the pipeline with kubeconform error "failed validation: could not find schema for CustomResourceDefinition".
Root cause is the naming of the file.
We copied the customresourcedefinitions.json from this repository and tried out with singular and plural naming. It seems that only singular naming is picked up by kubeconform.